relation v0.1.1 mix gen.json

Generates a Phoenix resource.

mix phoenix.gen.json User users name:string age:integer

The first argument is the module name followed by its plural name (used for resources and schema).

The generated resource will contain:

a schema in web/models

a view in web/views

a controller in web/controllers

a migration file for the repository

test files for generated model and controller

If you already have a model, the generated model can be skipped with --no-model. Read the documentation for phoenix.gen.model for more information on attributes and namespaced resources.

Functions

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.